Create list of required parameters for ProxyErrorSpectrum

Returns a list of arguments to ProxyErrorSpectrum where arguments in the call replace the defaults. Default values are returned for all arguments not in the call.

Usage

GetSpecPars(proxy.type, ...)

Arguments

proxy.type

"Mg_Ca", "T_deg_Mg_Ca", "T_deg_Uk37", "Uk37" or "d18O"

...

Arguments to ProxyErrorSpectrum, e.g. delta_t = 1000

Value

A list of arguments to ProxyErrorSpectrum

Examples

if (FALSE) {
GetSpecPars()
}
GetSpecPars(proxy.type = "Mg_Ca")
#> $delta_t
#> [1] 100
#> 
#> $tau_r
#> [1] 100
#> 
#> $tau_b
#> [1] 200
#> 
#> $tau_s
#> [1] 20
#> 
#> $T
#> [1] 10100
#> 
#> $tau_p
#> [1] 0.3333333
#> 
#> $N
#> [1] 30
#> 
#> $n.k
#> [1] 15
#> 
#> $clim.spec.fun
#> [1] "ClimPowerFunction"
#> 
#> $clim.spec.fun.args
#> $clim.spec.fun.args$a
#> [1] 0.1
#> 
#> $clim.spec.fun.args$b
#> [1] 1
#> 
#> 
#> $seas.amp
#> [1] 4
#> 
#> $sig.sq_a
#> [1] 0.004
#> 
#> $sig.sq_c
#> [1] 2
#> 
#> $phi_c
#> [1] 0
#> 
#> $delta_phi_c
#> [1] 0
#> 
#> $nu_a
#> [1] 4.347826e-05
#> 
#> $nu_c
#> [1] 1
#> 
#> $phi_a
#> [1] 1.570796
#> 
#> $sigma.meas
#> [1] 0.26
#> 
#> $sigma.ind
#> [1] 1.5
#> 
#> $sigma.cal
#> [1] 0.3
#> 
GetSpecPars(proxy.type = "Uk37")
#> $delta_t
#> [1] 100
#> 
#> $tau_r
#> [1] 100
#> 
#> $tau_b
#> [1] 200
#> 
#> $tau_s
#> [1] 20
#> 
#> $T
#> [1] 10100
#> 
#> $tau_p
#> [1] 0.6666667
#> 
#> $N
#> [1] Inf
#> 
#> $n.k
#> [1] 15
#> 
#> $clim.spec.fun
#> [1] "ClimPowerFunction"
#> 
#> $clim.spec.fun.args
#> $clim.spec.fun.args$a
#> [1] 0.1
#> 
#> $clim.spec.fun.args$b
#> [1] 1
#> 
#> 
#> $seas.amp
#> [1] 4
#> 
#> $sig.sq_a
#> [1] 0.004
#> 
#> $sig.sq_c
#> [1] 2
#> 
#> $phi_c
#> [1] 0
#> 
#> $delta_phi_c
#> [1] 0
#> 
#> $nu_a
#> [1] 4.347826e-05
#> 
#> $nu_c
#> [1] 1
#> 
#> $phi_a
#> [1] 1.570796
#> 
#> $sigma.meas
#> [1] 0.23
#> 
#> $sigma.ind
#> [1] 0
#> 
#> $sigma.cal
#> [1] 0.5
#> 
do.call(ProxyErrorSpectrum, GetSpecPars("Mg_Ca"))
#> Warning: Rounding T to 10100 so that T is an odd integer multiple of delta_t
#> $spec.pars
#> $spec.pars$nu
#> NULL
#> 
#> $spec.pars$tau_s
#> [1] 20
#> 
#> $spec.pars$tau_b
#> [1] 200
#> 
#> $spec.pars$tau_p
#> [1] 0.3333333
#> 
#> $spec.pars$tau_r
#> [1] 100
#> 
#> $spec.pars$T
#> [1] 10100
#> 
#> $spec.pars$delta_t
#> [1] 100
#> 
#> $spec.pars$N
#> [1] 30
#> 
#> $spec.pars$n.k
#> [1] 15
#> 
#> $spec.pars$clim.spec.fun
#> [1] "ClimPowerFunction"
#> 
#> $spec.pars$clim.spec.fun.args
#> $spec.pars$clim.spec.fun.args$a
#> [1] 0.1
#> 
#> $spec.pars$clim.spec.fun.args$b
#> [1] 1
#> 
#> 
#> $spec.pars$sig.sq_a
#> [1] 0.004
#> 
#> $spec.pars$sig.sq_c
#> [1] 2
#> 
#> $spec.pars$nu_a
#> [1] 4.347826e-05
#> 
#> $spec.pars$nu_c
#> [1] 1
#> 
#> $spec.pars$phi_a
#> [1] 1.570796
#> 
#> $spec.pars$phi_c
#> [1] 0
#> 
#> $spec.pars$delta_phi_c
#> [1] 0
#> 
#> $spec.pars$sigma.meas
#> [1] 0.26
#> 
#> $spec.pars$sigma.ind
#> [1] 1.5
#> 
#> $spec.pars$sigma.cal
#> [1] 0.3
#> 
#> $spec.pars$n.nu.prime
#> [1] 1000
#> 
#> $spec.pars$seas.amp
#> [1] 4
#> 
#> 
#> $proxy.error.spec
#> # A tibble: 101 × 12
#>          nu Climate Reference.climate Bioturbation Aliasing.stochastic
#>       <dbl>   <dbl>             <dbl>        <dbl>               <dbl>
#>  1 -0.00495    20.2              13.0         16.8                6.37
#>  2 -0.00485    20.6              13.5         17.0                6.37
#>  3 -0.00475    21.0              14.1         17.3                6.37
#>  4 -0.00465    21.5              14.6         17.8                6.37
#>  5 -0.00455    22.0              15.2         18.5                6.37
#>  6 -0.00446    22.4              15.8         19.3                6.37
#>  7 -0.00436    23.0              16.4         20.3                6.37
#>  8 -0.00426    23.5              17.1         21.4                6.37
#>  9 -0.00416    24.0              17.8         22.6                6.37
#> 10 -0.00406    24.6              18.5         23.9                6.37
#> # ℹ 91 more rows
#> # ℹ 7 more variables: Aliasing.seasonal <dbl>, Seasonal.bias <dbl>,
#> #   Seasonal.bias.unc. <dbl>, Meas.error <dbl>, Individual.variation <dbl>,
#> #   Calibration.unc. <dbl>, Total.error <dbl>
#> 
#> $Sign.SB
#> [1] 1
#> 
#> attr(,"class")
#> [1] "proxy.error.spec" "list"