Get the orbital error components

Get the orbital error components

Usage

OrbitalError(
  nu = NULL,
  tau_s,
  tau_b,
  tau_p,
  delta_t,
  T,
  sig.sq_c,
  phi_c,
  delta_phi_c,
  nu_c,
  sig.sq_a,
  phi_a,
  nu_a,
  N
)

Arguments

nu

frequency

tau_s

sediment slice thickness in years (layer.width / sedimentation rate)

tau_b

timescale of bioturbation (bioturbation depth / sedimentation rate) (L/sr)

tau_p

length of proxy carrier "growth season" (< 12 months)

delta_t

sampling frequency of the sediment core / climate timeseries

T

length of a finite time series, e.g. 1e04

sig.sq_c

variance of the seasonal cycle

phi_c

phase of carrier growth season relative to the maximum of the seasonal cycle

delta_phi_c

Uncertainty in the phase of the signal carrier production. A value between 0 and 2Pi.

nu_c

1/1 (yr) = frequency of the seasonal cycle

sig.sq_a

variance of precessionary amplitude modulation

phi_a

phase of precessionary amplitude modulation relative to centre of finite timeseries of length T

nu_a

1/tau_a = frequency of the orbital variation, e.g. for precession 1/23e03 yrs

N

number of signal carriers (e.g. individual foraminifera)

Value

A list with the frequency domain and time domain error components

Examples

spec.pars <- GetSpecPars("Mg_Ca", phi_c = pi, delta_phi_c = pi, sig.sq_a = 0.1, tau_p = 1/12)
do.call(OrbitalError, spec.pars[names(spec.pars) %in% names(formals(OrbitalError))])
#> $freq
#>                nu           SB          SU3       SU4
#> 1   -0.0049504950 3.016446e-01 7.212738e-02 0.1661818
#> 2   -0.0048514851 3.022294e-01 7.226719e-02 0.1661818
#> 3   -0.0047524752 3.034033e-01 7.254790e-02 0.1661818
#> 4   -0.0046534653 3.051757e-01 7.297170e-02 0.1661818
#> 5   -0.0045544554 3.075604e-01 7.354192e-02 0.1661818
#> 6   -0.0044554455 3.105764e-01 7.426309e-02 0.1661818
#> 7   -0.0043564356 3.142480e-01 7.514100e-02 0.1661818
#> 8   -0.0042574257 3.186051e-01 7.618285e-02 0.1661818
#> 9   -0.0041584158 3.236840e-01 7.739730e-02 0.1661818
#> 10  -0.0040594059 3.295281e-01 7.879468e-02 0.1661818
#> 11  -0.0039603960 3.361881e-01 8.038720e-02 0.1661818
#> 12  -0.0038613861 3.437240e-01 8.218913e-02 0.1661818
#> 13  -0.0037623762 3.522053e-01 8.421712e-02 0.1661818
#> 14  -0.0036633663 3.617129e-01 8.649053e-02 0.1661818
#> 15  -0.0035643564 3.723411e-01 8.903187e-02 0.1661818
#> 16  -0.0034653465 3.841991e-01 9.186728e-02 0.1661818
#> 17  -0.0033663366 3.974141e-01 9.502717e-02 0.1661818
#> 18  -0.0032673267 4.121342e-01 9.854695e-02 0.1661818
#> 19  -0.0031683168 4.285324e-01 1.024680e-01 0.1661818
#> 20  -0.0030693069 4.468112e-01 1.068387e-01 0.1661818
#> 21  -0.0029702970 4.672088e-01 1.117160e-01 0.1661818
#> 22  -0.0028712871 4.900061e-01 1.171672e-01 0.1661818
#> 23  -0.0027722772 5.155362e-01 1.232718e-01 0.1661818
#> 24  -0.0026732673 5.441962e-01 1.301248e-01 0.1661818
#> 25  -0.0025742574 5.764620e-01 1.378400e-01 0.1661818
#> 26  -0.0024752475 6.129074e-01 1.465546e-01 0.1661818
#> 27  -0.0023762376 6.542295e-01 1.564353e-01 0.1661818
#> 28  -0.0022772277 7.012810e-01 1.676859e-01 0.1661818
#> 29  -0.0021782178 7.551139e-01 1.805581e-01 0.1661818
#> 30  -0.0020792079 8.170377e-01 1.953649e-01 0.1661818
#> 31  -0.0019801980 8.886988e-01 2.125001e-01 0.1661818
#> 32  -0.0018811881 9.721892e-01 2.324638e-01 0.1661818
#> 33  -0.0017821782 1.070200e+00 2.558996e-01 0.1661818
#> 34  -0.0016831683 1.186240e+00 2.836463e-01 0.1661818
#> 35  -0.0015841584 1.324950e+00 3.168137e-01 0.1661818
#> 36  -0.0014851485 1.492574e+00 3.568950e-01 0.1661818
#> 37  -0.0013861386 1.697674e+00 4.059372e-01 0.1661818
#> 38  -0.0012871287 1.952239e+00 4.668073e-01 0.1661818
#> 39  -0.0011881188 2.273484e+00 5.436212e-01 0.1661818
#> 40  -0.0010891089 2.686834e+00 6.424590e-01 0.1661818
#> 41  -0.0009900990 3.231132e+00 7.726082e-01 0.1661818
#> 42  -0.0008910891 3.968135e+00 9.488356e-01 0.1661818
#> 43  -0.0007920792 5.000919e+00 1.195788e+00 0.1661818
#> 44  -0.0006930693 6.512217e+00 1.557160e+00 0.1661818
#> 45  -0.0005940594 8.851902e+00 2.116611e+00 0.1661818
#> 46  -0.0004950495 1.276192e+01 3.051549e+00 0.1661818
#> 47  -0.0003960396 2.005764e+01 4.796057e+00 0.1661818
#> 48  -0.0002970297 3.626242e+01 8.670843e+00 0.1661818
#> 49  -0.0001980198 8.610782e+01 2.058956e+01 0.1661818
#> 50  -0.0000990099 4.784555e+02 1.144053e+02 0.1661818
#> 51   0.0000000000 1.600228e+04 3.826365e+03 0.1661818
#> 52   0.0000990099 4.784555e+02 1.144053e+02 0.1661818
#> 53   0.0001980198 8.610782e+01 2.058956e+01 0.1661818
#> 54   0.0002970297 3.626242e+01 8.670843e+00 0.1661818
#> 55   0.0003960396 2.005764e+01 4.796057e+00 0.1661818
#> 56   0.0004950495 1.276192e+01 3.051549e+00 0.1661818
#> 57   0.0005940594 8.851902e+00 2.116611e+00 0.1661818
#> 58   0.0006930693 6.512217e+00 1.557160e+00 0.1661818
#> 59   0.0007920792 5.000919e+00 1.195788e+00 0.1661818
#> 60   0.0008910891 3.968135e+00 9.488356e-01 0.1661818
#> 61   0.0009900990 3.231132e+00 7.726082e-01 0.1661818
#> 62   0.0010891089 2.686834e+00 6.424590e-01 0.1661818
#> 63   0.0011881188 2.273484e+00 5.436212e-01 0.1661818
#> 64   0.0012871287 1.952239e+00 4.668073e-01 0.1661818
#> 65   0.0013861386 1.697674e+00 4.059372e-01 0.1661818
#> 66   0.0014851485 1.492574e+00 3.568950e-01 0.1661818
#> 67   0.0015841584 1.324950e+00 3.168137e-01 0.1661818
#> 68   0.0016831683 1.186240e+00 2.836463e-01 0.1661818
#> 69   0.0017821782 1.070200e+00 2.558996e-01 0.1661818
#> 70   0.0018811881 9.721892e-01 2.324638e-01 0.1661818
#> 71   0.0019801980 8.886988e-01 2.125001e-01 0.1661818
#> 72   0.0020792079 8.170377e-01 1.953649e-01 0.1661818
#> 73   0.0021782178 7.551139e-01 1.805581e-01 0.1661818
#> 74   0.0022772277 7.012810e-01 1.676859e-01 0.1661818
#> 75   0.0023762376 6.542295e-01 1.564353e-01 0.1661818
#> 76   0.0024752475 6.129074e-01 1.465546e-01 0.1661818
#> 77   0.0025742574 5.764620e-01 1.378400e-01 0.1661818
#> 78   0.0026732673 5.441962e-01 1.301248e-01 0.1661818
#> 79   0.0027722772 5.155362e-01 1.232718e-01 0.1661818
#> 80   0.0028712871 4.900061e-01 1.171672e-01 0.1661818
#> 81   0.0029702970 4.672088e-01 1.117160e-01 0.1661818
#> 82   0.0030693069 4.468112e-01 1.068387e-01 0.1661818
#> 83   0.0031683168 4.285324e-01 1.024680e-01 0.1661818
#> 84   0.0032673267 4.121342e-01 9.854695e-02 0.1661818
#> 85   0.0033663366 3.974141e-01 9.502717e-02 0.1661818
#> 86   0.0034653465 3.841991e-01 9.186728e-02 0.1661818
#> 87   0.0035643564 3.723411e-01 8.903187e-02 0.1661818
#> 88   0.0036633663 3.617129e-01 8.649053e-02 0.1661818
#> 89   0.0037623762 3.522053e-01 8.421712e-02 0.1661818
#> 90   0.0038613861 3.437240e-01 8.218913e-02 0.1661818
#> 91   0.0039603960 3.361881e-01 8.038720e-02 0.1661818
#> 92   0.0040594059 3.295281e-01 7.879468e-02 0.1661818
#> 93   0.0041584158 3.236840e-01 7.739730e-02 0.1661818
#> 94   0.0042574257 3.186051e-01 7.618285e-02 0.1661818
#> 95   0.0043564356 3.142480e-01 7.514100e-02 0.1661818
#> 96   0.0044554455 3.105764e-01 7.426309e-02 0.1661818
#> 97   0.0045544554 3.075604e-01 7.354192e-02 0.1661818
#> 98   0.0046534653 3.051757e-01 7.297170e-02 0.1661818
#> 99   0.0047524752 3.034033e-01 7.254790e-02 0.1661818
#> 100  0.0048514851 3.022294e-01 7.226719e-02 0.1661818
#> 101  0.0049504950 3.016446e-01 7.212738e-02 0.1661818
#> 
#> $time
#>       tax          B     U3.sq        U4.sq
#> 1   -5000 -1.8090709 0.7825565 0.0031182685
#> 2   -4900 -1.8057410 0.7796783 0.0031071333
#> 3   -4800 -1.8020029 0.7764536 0.0030946735
#> 4   -4700 -1.7978595 0.7728870 0.0030809062
#> 5   -4600 -1.7933137 0.7689836 0.0030658503
#> 6   -4500 -1.7883690 0.7647488 0.0030495262
#> 7   -4400 -1.7830291 0.7601886 0.0030319561
#> 8   -4300 -1.7772979 0.7553096 0.0030131639
#> 9   -4200 -1.7711798 0.7501184 0.0029931749
#> 10  -4100 -1.7646793 0.7446224 0.0029720158
#> 11  -4000 -1.7578012 0.7388292 0.0029497149
#> 12  -3900 -1.7505507 0.7327468 0.0029263020
#> 13  -3800 -1.7429332 0.7263836 0.0029018078
#> 14  -3700 -1.7349544 0.7197483 0.0028762647
#> 15  -3600 -1.7266202 0.7128501 0.0028497061
#> 16  -3500 -1.7179369 0.7056981 0.0028221664
#> 17  -3400 -1.7089109 0.6983022 0.0027936813
#> 18  -3300 -1.6995490 0.6906721 0.0027642873
#> 19  -3200 -1.6898582 0.6828181 0.0027340220
#> 20  -3100 -1.6798456 0.6747506 0.0027029235
#> 21  -3000 -1.6695188 0.6664800 0.0026710311
#> 22  -2900 -1.6588854 0.6580173 0.0026383844
#> 23  -2800 -1.6479535 0.6493733 0.0026050239
#> 24  -2700 -1.6367311 0.6405591 0.0025709903
#> 25  -2600 -1.6252266 0.6315858 0.0025363251
#> 26  -2500 -1.6134487 0.6224649 0.0025010699
#> 27  -2400 -1.6014061 0.6132075 0.0024652668
#> 28  -2300 -1.5891077 0.6038252 0.0024289579
#> 29  -2200 -1.5765629 0.5943293 0.0023921856
#> 30  -2100 -1.5637808 0.5847312 0.0023549924
#> 31  -2000 -1.5507712 0.5750425 0.0023174206
#> 32  -1900 -1.5375436 0.5652745 0.0022795126
#> 33  -1800 -1.5241079 0.5554385 0.0022413104
#> 34  -1700 -1.5104743 0.5455458 0.0022028560
#> 35  -1600 -1.4966528 0.5356075 0.0021641909
#> 36  -1500 -1.4826537 0.5256346 0.0021253564
#> 37  -1400 -1.4684876 0.5156382 0.0020863933
#> 38  -1300 -1.4541649 0.5056289 0.0020473417
#> 39  -1200 -1.4396964 0.4956172 0.0020082414
#> 40  -1100 -1.4250929 0.4856137 0.0019691314
#> 41  -1000 -1.4103653 0.4756284 0.0019300500
#> 42   -900 -1.3955245 0.4656713 0.0018910349
#> 43   -800 -1.3805816 0.4557521 0.0018521228
#> 44   -700 -1.3655478 0.4458804 0.0018133496
#> 45   -600 -1.3504343 0.4360652 0.0017747504
#> 46   -500 -1.3352524 0.4263157 0.0017363592
#> 47   -400 -1.3200134 0.4166402 0.0016982091
#> 48   -300 -1.3047287 0.4070474 0.0016603323
#> 49   -200 -1.2894097 0.3975451 0.0016227597
#> 50   -100 -1.2740677 0.3881410 0.0015855212
#> 51      0 -1.2587144 0.3788427 0.0015486456
#> 52    100 -1.2433611 0.3696571 0.0015121605
#> 53    200 -1.2280192 0.3605910 0.0014760926
#> 54    300 -1.2127003 0.3516507 0.0014404669
#> 55    400 -1.1974157 0.3428423 0.0014053077
#> 56    500 -1.1821769 0.3341716 0.0013706379
#> 57    600 -1.1669953 0.3256437 0.0013364790
#> 58    700 -1.1518821 0.3172638 0.0013028517
#> 59    800 -1.1368486 0.3090365 0.0012697751
#> 60    900 -1.1219061 0.3009661 0.0012372672
#> 61   1000 -1.1070658 0.2930565 0.0012053449
#> 62   1100 -1.0923386 0.2853114 0.0011740238
#> 63   1200 -1.0777356 0.2777340 0.0011433184
#> 64   1300 -1.0632677 0.2703272 0.0011132418
#> 65   1400 -1.0489456 0.2630937 0.0010838062
#> 66   1500 -1.0347801 0.2560358 0.0010550226
#> 67   1600 -1.0207818 0.2491554 0.0010269008
#> 68   1700 -1.0069610 0.2424543 0.0009994497
#> 69   1800 -0.9933281 0.2359337 0.0009726769
#> 70   1900 -0.9798933 0.2295948 0.0009465893
#> 71   2000 -0.9666665 0.2234384 0.0009211926
#> 72   2100 -0.9536577 0.2174651 0.0008964916
#> 73   2200 -0.9408766 0.2116752 0.0008724901
#> 74   2300 -0.9283327 0.2060686 0.0008491912
#> 75   2400 -0.9160354 0.2006453 0.0008265971
#> 76   2500 -0.9039938 0.1954049 0.0008047093
#> 77   2600 -0.8922169 0.1903467 0.0007835285
#> 78   2700 -0.8807135 0.1854701 0.0007630545
#> 79   2800 -0.8694923 0.1807740 0.0007432869
#> 80   2900 -0.8585615 0.1762574 0.0007242244
#> 81   3000 -0.8479294 0.1719190 0.0007058651
#> 82   3100 -0.8376038 0.1677574 0.0006882069
#> 83   3200 -0.8275925 0.1637712 0.0006712470
#> 84   3300 -0.8179029 0.1599588 0.0006549823
#> 85   3400 -0.8085423 0.1563184 0.0006394092
#> 86   3500 -0.7995177 0.1528483 0.0006245242
#> 87   3600 -0.7908357 0.1495468 0.0006103231
#> 88   3700 -0.7825030 0.1464119 0.0005968017
#> 89   3800 -0.7745256 0.1434419 0.0005839557
#> 90   3900 -0.7669095 0.1406348 0.0005717806
#> 91   4000 -0.7596605 0.1379887 0.0005602719
#> 92   4100 -0.7527839 0.1355019 0.0005494249
#> 93   4200 -0.7462849 0.1331723 0.0005392351
#> 94   4300 -0.7401683 0.1309983 0.0005296982
#> 95   4400 -0.7344387 0.1289780 0.0005208096
#> 96   4500 -0.7291004 0.1271098 0.0005125653
#> 97   4600 -0.7241573 0.1253921 0.0005049611
#> 98   4700 -0.7196131 0.1238234 0.0004979932
#> 99   4800 -0.7154712 0.1224021 0.0004916580
#> 100  4900 -0.7117347 0.1211270 0.0004859521
#> 101  5000 -0.7084065 0.1199968 0.0004808727
#>