Get error components for stochastic climate signal

Get error components for stochastic climate signal

Usage

S_E(
  nu,
  tau_s,
  tau_b,
  tau_r,
  T,
  delta_t,
  N,
  n.k,
  clim.spec.fun,
  clim.spec.fun.args,
  n.nu.prime = 1000,
  plot.integral = FALSE
)

Arguments

nu

frequency

tau_s

sediment slice thickness in years (layer.width / sedimentation rate)

tau_b

timescale of bioturbation (bioturbation depth / sedimentation rate) (L/sr)

tau_r

width of moving average filter that represents the "interpreted" resolution of the timeseries

T

length of a finite time series, e.g. 1e04

delta_t

sampling frequency of the sediment core / climate timeseries

N

number of signal carriers (e.g. individual foraminifera)

n.k

the number of aliasers used when estimating the error spectrum of the stochastic climate signal. Defaults to 15, does not normally need to be adjusted.

clim.spec.fun

a function to return climate power spectral density as a function of frequency, nu

clim.spec.fun.args

arguments to the named climate power spectrum function

n.nu.prime

number of discrete frequencies at which to evaluate the PSD of the error

plot.integral

Plot the PSD of the error of the bioturbated climate timeseries. This PSD is numerically integrated to give the variance for the aliasing of climate variation.

Value

list of frequencies and error spectrum components

See also

Examples

spec.pars <- GetSpecPars("Mg_Ca", phi_c = pi, delta_phi_c = pi,
 sig.sq_a = 0.1, tau_p = 1/12)
spec.pars$nu <- GetNu(spec.pars$T, spec.pars$delta_t)
do.call(S_E, spec.pars[names(spec.pars) %in% names(formals(S_E))])
#> $nu
#>   [1] -0.0049504950 -0.0048514851 -0.0047524752 -0.0046534653 -0.0045544554
#>   [6] -0.0044554455 -0.0043564356 -0.0042574257 -0.0041584158 -0.0040594059
#>  [11] -0.0039603960 -0.0038613861 -0.0037623762 -0.0036633663 -0.0035643564
#>  [16] -0.0034653465 -0.0033663366 -0.0032673267 -0.0031683168 -0.0030693069
#>  [21] -0.0029702970 -0.0028712871 -0.0027722772 -0.0026732673 -0.0025742574
#>  [26] -0.0024752475 -0.0023762376 -0.0022772277 -0.0021782178 -0.0020792079
#>  [31] -0.0019801980 -0.0018811881 -0.0017821782 -0.0016831683 -0.0015841584
#>  [36] -0.0014851485 -0.0013861386 -0.0012871287 -0.0011881188 -0.0010891089
#>  [41] -0.0009900990 -0.0008910891 -0.0007920792 -0.0006930693 -0.0005940594
#>  [46] -0.0004950495 -0.0003960396 -0.0002970297 -0.0001980198 -0.0000990099
#>  [51]  0.0000000000  0.0000990099  0.0001980198  0.0002970297  0.0003960396
#>  [56]  0.0004950495  0.0005940594  0.0006930693  0.0007920792  0.0008910891
#>  [61]  0.0009900990  0.0010891089  0.0011881188  0.0012871287  0.0013861386
#>  [66]  0.0014851485  0.0015841584  0.0016831683  0.0017821782  0.0018811881
#>  [71]  0.0019801980  0.0020792079  0.0021782178  0.0022772277  0.0023762376
#>  [76]  0.0024752475  0.0025742574  0.0026732673  0.0027722772  0.0028712871
#>  [81]  0.0029702970  0.0030693069  0.0031683168  0.0032673267  0.0033663366
#>  [86]  0.0034653465  0.0035643564  0.0036633663  0.0037623762  0.0038613861
#>  [91]  0.0039603960  0.0040594059  0.0041584158  0.0042574257  0.0043564356
#>  [96]  0.0044554455  0.0045544554  0.0046534653  0.0047524752  0.0048514851
#> [101]  0.0049504950
#> 
#> $inf.N.part
#>   [1] 16.8306892 17.0020892 17.3430234 17.8497763 18.5168086 19.3367981
#>   [7] 20.3006942 21.3977849 22.6157769 23.9408880 25.3579515 26.8505320
#>  [13] 28.4010520 29.9909285 31.6007186 33.2102743 34.7989040 36.3455407
#>  [19] 37.8289156 39.2277357 40.5208647 41.6875056 42.7073841 43.5609317
#>  [25] 44.2294686 44.6953846 44.9423195 44.9553438 44.7211408 44.2281966
#>  [31] 43.4670020 42.4302783 41.1132419 39.5139306 37.6336256 35.4774158
#>  [37] 33.0549728 30.3816251 27.4798487 24.3813075 21.1295661 17.7835060
#>  [43] 14.4212056 11.1434339  8.0747606  5.3585480  3.1403999  1.5351555
#>  [49]  0.5790584  0.1833695        NaN  0.1833695  0.5790584  1.5351555
#>  [55]  3.1403999  5.3585480  8.0747606 11.1434339 14.4212056 17.7835060
#>  [61] 21.1295661 24.3813075 27.4798487 30.3816251 33.0549728 35.4774158
#>  [67] 37.6336256 39.5139306 41.1132419 42.4302783 43.4670020 44.2281966
#>  [73] 44.7211408 44.9553438 44.9423195 44.6953846 44.2294686 43.5609317
#>  [79] 42.7073841 41.6875056 40.5208647 39.2277357 37.8289156 36.3455407
#>  [85] 34.7989040 33.2102743 31.6007186 29.9909285 28.4010520 26.8505320
#>  [91] 25.3579515 23.9408880 22.6157769 21.3977849 20.3006942 19.3367981
#>  [97] 18.5168086 17.8497763 17.3430234 17.0020892 16.8306892
#> 
#> $finite.N.part
#> [1] 6.37409
#>